User Sign Up Logic and index.js adjustment
This commit features successful user sign up functions.
The submit button on the form initiates the function, the function runs
out of index.js, which requires events.js, which requires api.js and
ui.js

This commit also makes a subtle change in index.js - The event listener
for sign-out is now `submit`, not `click` because it is nested in a
form. This requires further consideration: Does a submit button need a
form? Will test various ways soon while making the sign-out logic.

Next up: User sign-in logic