Tie functionality, beginning of game onReset
This commit features the ability to determine if there was a tie in the
game. This commit also features a new function called `onReset` which
will be used if a winner or tie was established. `onReset` will halt
gameplay (no further spaces can be played) and display a game outcome
message ("winner is X/O" or "game resulted in a tie") and then offer
the users a button to reset the game. Selecting reset will set all
gloabl variables back to their original state and clear the board, user
X will go first again.

Up next: build out onReset