View State Hidden Game
This commit features a new view state control. The view state changes
during onSignIn and onSignOut, and the change occurs by adding or
removing the bootstrap class `hidden` from the gb-section in the html,
which shows or hides the entire game. The game is visible when signed
in and hidden when signed out.