Board clear on sign-out bug
Bug:
```md
Do Not
Have any user-facing bugs.

Board does not clear when signing out. Second user can sign in and
continue playing previous game.
```

This user facing bug was not clearing the gameboard when a user signed
out - if user a plays half of a game, then signs out, and user b signs
in, user a's half-played game will appear for user b.

This commit resolves this bug by requiring engine.js onReset function
through authentication ui.js and invoking onReset during sign in
success and sign out success (obviously not needed for a sign in error,
and it seems like bad ui to clear the board on a sign-out error), just
to be certain the game board clears.